Replace docs.tsx with corpus.tsx (keeping the former name)
