+++ packages/ai-jsx/src/batteries/docs.tsx
@@ -1,156 +1,366 @@
-import { Document as LangChainDocument } from 'langchain/document';
-import { Embeddings } from 'langchain/embeddings/base';
-import { TokenTextSplitter } from 'langchain/text_splitter';
-import { MemoryVectorStore } from 'langchain/vectorstores/memory';
+import { similarity } from "ml-distance";
 import { Jsonifiable } from 'type-fest';
 import { ChatCompletion, SystemMessage, UserMessage } from '../core/completion';
 import * as LLMx from '../index.js';
 import { Node } from '../index.js';
 
-// TODO: Finish replacing this with corpus.tsx.
-
 /**
- * A document that can be queried.
+ * A raw document loaded from an arbitrary source that has not yet been parsed.
  */
-export interface Document<Metadata extends Jsonifiable = Jsonifiable> {
-  /**
-   * The content of the document.
+export class RawDocument {
+  constructor(
+    /** The raw bytes content. */
+    readonly pageContent: Uint8Array,
+    /**
+   * A human-readable name for this document. This identifies the document to users and should
+   * be unique within a corpus. If the document is hosted on the web, its URL is a good choice.
    */
-  pageContent: string;
-
-  /**
-   * Metadata about the document, such as the URL at which it's hosted, the title, or when it was created.
-   */
-  metadata?: Metadata;
+    readonly name?: string,
+    /** The encoding used for this document. Defaults to utf-8. */
+    readonly encoding: string = 'utf-8',
+    /** The content type for this document. Defaults to plain text. */
+    readonly mimeType: string = 'text/plain') {
+      this.pageContent = pageContent;
+      this.name = name;
+      this.encoding = encoding;
+      this.mimeType = mimeType;
+    }
 }
 
 /**
- * Chunk text into multiple documents.
- *
- * In LangChain, this is called "split". We use the phrase "chunk", because the cannonical term for the output
- * of this process is "chunks".
+ * A text document that can be used to ground responses.
  */
-export type Chunker = <Metadata extends Jsonifiable = Jsonifiable>(
-  doc: Document<Metadata>,
-  ...args: any[]
-) => Promise<Document<Metadata>[]>;
-export type ChunkMany = <Metadata extends Jsonifiable = Jsonifiable>(
-  doc: Document<Metadata>[],
-  ...args: any[]
-) => Promise<Document<Metadata>[]>;
+export class Document<DocumentMetadata extends Jsonifiable = Jsonifiable> {
+  constructor(
+    /**
+   * The content of the document. While this may often be a singleton, including separate
+   * meaningful pieces of text can aid in chunking/embedding.
+   */
+    readonly pageContent: string[],
+    /**
+   * A human-readable name for this document. This identifies the document to users and should
+   * be unique within a corpus. If the document is hosted on the web, its URL is a good choice.
+   */
+    readonly name?: string,
+    /**
+   * Other metadata about the document, such as when it was created.
+   */
+    readonly metadata?: DocumentMetadata) {
+      this.pageContent = pageContent;
+      this.name = name;
+      this.metadata = metadata;
+    }
+}
 
 /**
- * Load Documents from a source.
- *
- * Examples of when you'd use this:
- *  * Load and parse a file into documents.
- *  * Call a remote API
- *  * Load data from a database
- *  * Do a k-nearest-neighbors search in a vector store.
- *
- * Once you have the Documents, you may wish to do some combination of:
- *  1. Inserting them directly into a prompt
- *  2. Adding them to a vector store.
- *  3. Chunking them into smaller documents (use a `Chunk` method).
- *
- * In LangChain, there are loaders and retrievers. In AI.JSX, there are only loaders. Retrievers are a special
- * case of loaders.
+ * A function that parses raw documents to produce text documents that can be used by an LLM.
  */
-// TODO: maybe rename to DocumentProvider
-export type Loader = (...args: any[]) => Promise<Document[]>;
+export type Parser<DocumentMetadata extends Jsonifiable = Jsonifiable> = (raw: RawDocument) => Promise<Document<DocumentMetadata>>;
 
-export interface VectorSearchResult<Metadata extends Jsonifiable = Jsonifiable> {
-  document: Document<Metadata>;
+function defaultParser<DocumentMetadata extends Jsonifiable = Jsonifiable>(raw: RawDocument): Promise<Document<DocumentMetadata>> {
+  if (raw.mimeType.startsWith('text/')) {
+    let content = new TextDecoder(raw.encoding).decode(raw.pageContent);
+    return Promise.resolve(new Document([content], raw.name));
+    // TODO: Add support for other mime types.
+  } else {
+    throw new Error(`Unsupported mime type: ${raw.mimeType}`);
+  }
+}
+
+/**
+ * Objects related to loading a corpus. This is facilitated by a series of
+ * requests and corresponding responses.
+ *
+ * In addition to returning documents, each response may expand the corpus
+ * space in one or both of two dimensions:
+ *      Responses may include new partitions to be loaded. Partitions are
+ *          non-overlapping subsets of a corpus which may be loaded in parallel.
+ *          A response's new partitions will be ignored if previously included
+ *          in another response.
+ *      When a response includes a page of documents, that page may indicate
+ *          that another page is available in the same partition. Pages are
+ *          always loaded serially in order. The partition is completed when
+ *          a response has a page with no next page token.
+ *
+ *  Loading will always begin with a  request with the default (unnamed)
+ *  partition and no page token. Subsequent requests depend on prior responses
+ *  and will always include at least one of those fields.
+ *
+ *  Examples:
+ *      Simple handful of documents:
+ *          The response to the initial request contains only a page of
+ *          documents. This could include a next_page_token for more
+ *          documents in the single default partition if needed.
+ *      Web crawl:
+ *          Each URL corresponds to a partition and the responses never include
+ *          tokens. The initial response only includes partitions, one for each
+ *          root URL to crawl. Each subsequent request includes the partition
+ *          (the URL) and the corresponding response contains a page with a
+ *          single document - the resource at that URL. If the document links
+ *          to other resources that should be included in the corpus, the
+ *          response also contains those URLs as new partitions. The process
+ *          repeats for all partitions until there are no known incomplete
+ *          partitions or until crawl limits are reached.
+ *      Database:
+ *          Consider a database with a parent table keyed by parent_id and an
+ *          interleaved child table keyed by (parent_id, child_id) whose rows
+ *          correspond to corpus documents. This loader will use tokens that
+ *          encode a read timestamp (for consistency) and an offset to be used
+ *          in combination with a static page size.
+ *
+ *          Upon receiving the initial request, the loader chooses a commit
+ *          timestamp to use for all reads and returns a partition for each
+ *          parent_id along with a first page token indicating the chosen read
+ *          timestamp and an offset of 0.
+ *
+ *          For each partition, the loader then receives requests with the
+ *          partition (a parent_id) and a page token (the read timestamp and
+ *          offest). It responds with documents corresponding to the next page
+ *          size child rows within the given parent. If more children exist,
+ *          the response includes a next page token with the same read
+ *          timestamp and an incremented offset. This repeats until there are
+ *          no more children, at which point the response has no
+ *          next page token and the partition is complete.
+ *
+ *          Note: Including multiple parent_ids in each partition would also
+ *              work and would be an effective way to limit parallelism if
+ *              desired.
+ */
+export namespace CorpusLoading {
+  export class Request {
+    constructor(readonly partition?: string, readonly pageToken?: string) {
+      /**
+       * The partition of the corpus that should be read. This will
+       * be empty for the initial request, indicating the default partition.
+       * For subsequent requests, it will typically be the name of a partition
+       * returned by a previous request, though it could be empty if the
+       * default partition contains multiple pages.
+       */
+      this.partition = partition;
+      /**
+       * A token for paginating results within a corpus partition.
+       * If present, this will be echoed from a previous response.
+       */
+      this.pageToken = pageToken;
+    }
+  }
+
+  export class Response<DocType extends RawDocument | Document> {
+    constructor(readonly page?: Page<DocType>, readonly partitions?: Partition[]) {
+      /** A page of documents from the requested partition. */
+      this.page = page;
+      /**
+       * Additional partitions that should be loaded in subsequent requests.
+       * These will be ignored if they were already included in a previous
+       * response.
+       */
+      this.partitions = partitions;
+    }
+  }
+
+  export class Page<DocType extends RawDocument | Document> {
+    constructor(readonly documents: DocType[], readonly nextPageToken?: string) {
+      /** Documents loaded into the corpus. */
+      this.documents = documents;
+      /**
+       * If there's another page in the requested partition, the token needed
+       * for requesting it.
+       */
+      this.nextPageToken = nextPageToken;
+    }
+  }
+
+  export class Partition {
+    constructor(readonly name: string, readonly firstPageToken?: string) {
+      /** The unique name of this partition. */
+      this.name = name;
+      /** An optional token for requesting the first page. */
+      this.firstPageToken = firstPageToken;
+    }
+  }
+
+  /** A function responsible for loading and parsing a corpus. */
+  export type Loader<DocumentMetadata extends Jsonifiable = Jsonifiable> = (request: Request) => Promise<Response<Document<DocumentMetadata>>>;
+
+  /** A function responsible for loading a corpus when parsing is handled separately. */
+  export type RawLoader = (request: Request) => Promise<Response<RawDocument>>;
+
+  export function toLoader<DocumentMetadata extends Jsonifiable = Jsonifiable>(rawLoader: RawLoader, parser: Parser<DocumentMetadata> = defaultParser): Loader<DocumentMetadata> {
+    return async (request: Request) => {
+      const response = await rawLoader(request);
+      if (response.page) {
+        const documents = await Promise.all(response.page.documents.map(parser));
+        return new Response(new Page(documents, response?.page?.nextPageToken), response.partitions);
+      }
+      return new Response<Document<DocumentMetadata>>(undefined, response.partitions);
+    };
+  }
+}
+
+/** A function that splits a document into multiple chunks, for example to ensure it fits in appropriate context windows. */
+export type Chunker<DocumentMetadata extends Jsonifiable = Jsonifiable, ChunkMetadata extends Jsonifiable = Jsonifiable> = (document: Document<DocumentMetadata>) => Promise<Chunk<DocumentMetadata>[]>;
+
+/** A chunk is a piece of a document that's appropriately sized for an LLM's context window and for semantic search. */
+export interface Chunk<ChunkMetadata extends Jsonifiable = Jsonifiable> {
+  /** The content of this chunk. This is what is provided as context to an LLM when the chunk is selected. */
+  content: string;
+
+  /** The name of the document from which this chunk was extracted. */
+  documentName?: string;
+
+  /** Optional additional metadata associated with this chunk. */
+  metadata?: ChunkMetadata;
+}
+
+/**
+ * An embedding is a function that maps a string to a vector encoding its semantic meaning.
+ * Often this is based on the same function used to transform text for an LLM's transformers.
+ */
+export type Embedding = (text: string) => Promise<number[]>;
+
+/**
+ * An embedded chunk is a piece of a document that is ready to be added into a vector space.
+ */
+export interface EmbeddedChunk<ChunkMetadata extends Jsonifiable = Jsonifiable> {
+  /** The identifier for this chunk. For insertions, this will be generated if not populated.*/
+  id?: string;
+
+  /** The vector representing this chunk for semantic nearest neighbors. */
+  vector: number[];
+
+  /** The content of this chunk. This is what is provided as context to an LLM when the chunk is selected. */
+  content: string;
+
+  /** The name of the document from which this chunk was extracted. */
+  documentName?: string;
+
+  /** Optional additional metadata associated with this chunk. */
+  metadata?: ChunkMetadata;
+}
+
+/** A chunk along with its score for a particular query. */
+export interface ScoredChunk<ChunkMetadata extends Jsonifiable = Jsonifiable> {
+  chunk: EmbeddedChunk<ChunkMetadata>;
   score: number;
 }
 
-/**
- * This is a simplified version of LangChain's `VectorStore`.
- */
-export interface VectorStore {
-  // Hmm. If we have this return value, then VectorStore['search'] does not extend `Load`.
-  // If we wanted to fix this, we'd need a both `search` and `searchWithScores` method.
-  // We could also have a `loadFromVectorStore` method.
-  //    const loadFromVectorStore = (vs: VectorStore, query, opts) => _.map(vs.search(query, opts), 'document');
+export interface Corpus<ChunkMetadata extends Jsonifiable = Jsonifiable> {
+  startCrawl: () => Promise<Corpus.Stats>;
 
-  search<Filter = unknown>(query: string, opts: { filter?: Filter; limit: number }): Promise<VectorSearchResult[]>;
+  getStats: () => Promise<Corpus.Stats>;
+
+  search: (query: string, params?: {limit?: number, score_threshold?: number}) => Promise<ScoredChunk<ChunkMetadata>[]>;
 }
 
-/**
- * A vector database that stores vectors in memory. In-memory is nice for rapid prototyping. However, it'll be slower, because every run needs to populate the database. And it's limited by the current process' memory.
- */
-export class DefaultInMemoryVectorStore<Metadata extends Jsonifiable = Jsonifiable> implements VectorStore {
-  private constructor(private readonly langChainVectorStore: MemoryVectorStore) {}
-
-  static async fromDocuments<Metadata extends Jsonifiable = Jsonifiable>(
-    docs: Document<Metadata>[],
-    embeddings: Embeddings
-  ) {
-    const lcMemoryStore = await MemoryVectorStore.fromDocuments(
-      docs.map((doc) => toLangChainDoc(doc)),
-      embeddings
-    );
-    return new this<Metadata>(lcMemoryStore);
+export namespace Corpus {
+  export enum LoadingState {
+    NOT_STARTED = 'NOT_STARTED',
+    IN_PROGRESS = 'IN_PROGRESS',
+    COMPLETED = 'COMPLETED',
+    FAILED = 'FAILED',
   }
 
-  async search<Filter = never>(
-    query: string,
-    opts: { filter?: Filter | undefined; limit: number }
-  ): Promise<VectorSearchResult<Metadata>[]> {
-    const docs = await this.langChainVectorStore.similaritySearchWithScore(query, opts.limit);
-    return docs.map(([doc, score]) => ({
-      document: doc as Document<Metadata>,
-      score,
+  export interface Stats {
+    loadingState: LoadingState;
+    completedPartitions: number;
+    activePartitions: number;
+    numDocuments: number;
+    numChunks: number;
+  }
+}
+
+export class LocalCorpus<DocumentMetadata extends Jsonifiable = Jsonifiable, ChunkMetadata extends Jsonifiable = Jsonifiable> implements Corpus<ChunkMetadata> {
+  private readonly vectors: EmbeddedChunk<ChunkMetadata>[] = [];
+  private readonly documents: Document<DocumentMetadata>[] = [];
+  private readonly completedPartitions = new Set<string>();
+  private readonly activePartitionsToToken = new Map<string|null, string|null>();
+  private loadingState = Corpus.LoadingState.NOT_STARTED;
+  constructor(readonly loader: CorpusLoading.Loader<DocumentMetadata>,
+    readonly chunker: Chunker<DocumentMetadata, ChunkMetadata>,
+    readonly embedding: Embedding) {
+    this.loader = loader;
+    this.chunker = chunker;
+    this.embedding = embedding;
+  }
+
+  async startCrawl(): Promise<Corpus.Stats> {
+    if (this.loadingState !== Corpus.LoadingState.NOT_STARTED) {
+      return this.getStats();
+    }
+    this.loadingState = Corpus.LoadingState.IN_PROGRESS;
+    try {
+      await this.doCrawl();
+      this.loadingState = Corpus.LoadingState.COMPLETED;
+    } catch (e) {
+      this.loadingState = Corpus.LoadingState.FAILED;
+      throw e;
+    }
+    return this.getStats();
+  }
+
+  private async doCrawl(): Promise<void> {
+    this.activePartitionsToToken.set(null, null);
+    while (this.activePartitionsToToken.size > 0) {
+      const partition = this.activePartitionsToToken.keys().next().value;
+      const token = this.activePartitionsToToken.get(partition);
+      this.activePartitionsToToken.delete(partition);
+
+      const response = await this.loader({partition, token} as CorpusLoading.Request);
+      for (let newPartition of response.partitions ?? []) {
+        if (!this.completedPartitions.has(newPartition.name) && !this.activePartitionsToToken.has(newPartition.name)) {
+          this.activePartitionsToToken.set(newPartition.name, newPartition.firstPageToken ?? null);
+        }
+      }
+      if (response.page) {
+        this.documents.push(...response.page.documents);
+        const vectors = await Promise.all(response.page.documents.map(this.vectorize));
+        this.vectors.push(...vectors.flat());
+        if (response.page.nextPageToken) {
+          this.activePartitionsToToken.set(partition, response.page.nextPageToken);
+        } else {
+          this.completedPartitions.add(partition);
+        }
+      }
+      if (!response.page && ! response.partitions) {
+        // Corner case - invalid response.
+        this.completedPartitions.add(partition);
+      }
+    }
+  }
+
+  private async vectorize(doc: Document<DocumentMetadata>): Promise<EmbeddedChunk<ChunkMetadata>[]> {
+    const chunks = await this.chunker(doc);
+    return Promise.all(chunks.map(async chunk => {
+      const vector = await this.embedding(chunk.content);
+      return {vector, ...chunk} as EmbeddedChunk<ChunkMetadata>;
     }));
   }
-}
 
-/**
- * If you need to chunk your documents, start with this chunker. Move to more powerful chunkers only as needed.
- */
-export const defaultChunker = <Metadata extends Jsonifiable = Jsonifiable>(
-  doc: Document<Metadata>,
-  opts: ConstructorParameters<typeof TokenTextSplitter>
-) => {
-  const splitter = new TokenTextSplitter({
-    encodingName: 'gpt2',
-    chunkSize: 600,
-    chunkOverlap: 100,
-    ...opts,
-  });
+  async getStats(): Promise<Corpus.Stats> {
+    return {
+      loadingState: this.loadingState,
+      completedPartitions: this.completedPartitions.size,
+      activePartitions: this.activePartitionsToToken.size,
+      numDocuments: this.vectors.length,
+      numChunks: this.vectors.length,
+    } as Corpus.Stats;
+  }
 
-  // This is complaining because {} doesn't necessary satisfy Metadata, since Metadata could have required fields.
-  // However, I think it's unlikely to cause a problem in practice at this point.
-  // @ts-expect-error
-  doc.metadata = doc.metadata ?? {};
+  async search(query: string, params?: {limit?: number, score_threshold?: number}): Promise<ScoredChunk<ChunkMetadata>[]> {
+    const queryVector = await this.embedding(query);
+    return this.vectors.map((vector) => ({
+      chunk: vector,
+      score: similarity.cosine(queryVector, vector.vector),
+    } as ScoredChunk<ChunkMetadata>)).sort((a, b) => b.score - a.score).slice(0, params?.limit ?? 10);
+  }
 
-  return splitter.splitDocuments([
-    // TS doesn't realize that we've ensured that the metadata is defined.
-    // @ts-expect-error
-    doc,
-  ]);
-};
-
-/**
- * If you need to chunk your documents, start with this chunker. Move to more powerful chunkers only as needed.
- */
-export const defaultChunkMany = async <Metadata extends Jsonifiable = Jsonifiable>(
-  docs: Document<Metadata>[],
-  opts?: Parameters<ChunkMany>[1]
-) => (await Promise.all(docs.map((doc) => defaultChunker(doc, opts)))).flat();
-
-/**
- * A cast wrapper to convert an AI.JSX document to a LangChain document.
- */
-export function toLangChainDoc(doc: Document): LangChainDocument {
-  return doc as LangChainDocument;
 }
 
 export interface DocsQAProps<Doc extends Document> {
   /**
-   * A loader that returns a list of documents relevant to a query.
+   * The corpus of documents that may be relevant to a query.
    */
-  loader: Loader;
+  corpus: Corpus;
 
   /**
    * The query to answer.
@@ -175,18 +385,20 @@ export interface DocsQAProps<Doc extends Document> {
  * A component that can be used to answer questions about documents. This is a very common usecase for LLMs.
  */
 export async function DocsQA<Doc extends Document>(props: DocsQAProps<Doc>) {
-  const docs = await props.loader();
+  const docs = await props.corpus.search(props.question);
   return (
     <ChatCompletion>
-      <SystemMessage>
-        You are a customer service agent. Answer questions truthfully. Here is what you know:
-        {docs.map((doc) => (
-          // TODO improve types
-          // @ts-expect-error
-          <props.docComponent doc={doc} />
-        ))}
-      </SystemMessage>
-      <UserMessage>{props.question}</UserMessage>
+    <SystemMessage>
+    You are a customer service agent.Answer questions truthfully.Here is what you know:
+  {
+    docs.map((doc) => (
+      // TODO improve types
+      // @ts-expect-error
+      <props.docComponent doc= { doc } />
+        ))
+  }
+  </SystemMessage>
+    <UserMessage> { props.question } </UserMessage>
     </ChatCompletion>
   );
 }
